refactor: remove duplicated code from `.withIdPrefix()`
