+++ index.js
@@ -14,15 +14,15 @@ function hoodieApi (options) {
 
   return {
     db: this,
-    add: require('./add').bind(this),
-    find: require('./find').bind(this),
-    findAll: require('./find-all').bind(this),
-    findOrAdd: require('./find-or-add').bind(this),
-    update: require('./update').bind(this),
-    updateOrAdd: require('./update-or-add').bind(this),
-    updateAll: require('./update-all').bind(this),
-    remove: require('./remove').bind(this),
-    removeAll: require('./remove-all').bind(this),
+    add: require('./add').bind(this, null),
+    find: require('./find').bind(this, null),
+    findAll: require('./find-all').bind(this, null),
+    findOrAdd: require('./find-or-add').bind(this, state, null),
+    update: require('./update').bind(this, null),
+    updateOrAdd: require('./update-or-add').bind(this, null),
+    updateAll: require('./update-all').bind(this, null),
+    remove: require('./remove').bind(this, null),
+    removeAll: require('./remove-all').bind(this, null),
     withIdPrefix: require('./with-id-prefix').bind(this, state),
     on: require('./lib/on').bind(this, state),
     one: require('./lib/one').bind(this, state),
