+++ update.js
@@ -8,12 +8,13 @@ var updateMany = require('./helpers/update-many')
 /**
  * updates existing object.
  *
+ * @param  {String}          prefix       optional id prefix
  * @param  {String|Object}   idOrObject   id or object with `.id` property
  * @param  {Object|Function} [change]     Changed properties or function
  *                                        that changes existing object
  * @return {Promise}
  */
-function update (objectsOrIds, change) {
+function update (prefix, objectsOrIds, change) {
   if (typeof objectsOrIds !== 'object' && !change) {
     return Promise.reject(
       new Error('Must provide change')
@@ -21,6 +22,6 @@ function update (objectsOrIds, change) {
   }
 
   return Array.isArray(objectsOrIds)
-    ? updateMany.call(this, objectsOrIds, change)
-    : updateOne.call(this, objectsOrIds, change)
+    ? updateMany.call(this, objectsOrIds, change, prefix)
+    : updateOne.call(this, objectsOrIds, change, prefix)
 }
